Fixing some linker errors for gstrtsp.
Note: The conversion wizard seems to have lost a bunch of project dependencies.
If the linker complains about unknown symbols, most probably the problem is that
some dependant libraries are not included in the compilation.

How to fix this:

1. Check the Build/Windows/.../obj/(project).log to see what .libs are included.

2. Figure out what libraries might contain the missing symbols. Often this is quite
easy based on the function name, sometimes you need to use Google or grep.

3. Right-click on the project in Visual Studio and select References. Add the dependency
there.